Update chuck_util_functions.py
added midi info checker.  if you're on a line with `MidiIn` , then ctrl+shift+o  (or whatever you bind to file opener, or even right click the context works..)  boom. check the console. I may redirect this to the open window that chuck uses to output stdout. that's for later.